[AIP-21] Move sensors.metastore_partition_sensor providers.apache.hive.sensors.metastore_partition
